+++ eng/tools/azure-sdk-tools/azpysdk/pylint.py
@@ -7,12 +7,14 @@ from subprocess import CalledProcessError, check_call
 
 from .Check import Check
 from ci_tools.functions import install_into_venv
-from ci_tools.variables import discover_repo_root, in_ci, set_envvar_defaults, in_ci, set_envvar_defaults
+from ci_tools.scenario.generation import create_package_and_install
+from ci_tools.variables import discover_repo_root, in_ci, set_envvar_defaults
 from ci_tools.environment_exclusions import is_check_enabled
 from ci_tools.logging import logger
 
 REPO_ROOT = discover_repo_root()
 PYLINT_VERSION = "3.2.7"
+PYGITHUB_VERSION = "1.59.0"
 
 class pylint(Check):
     def __init__(self) -> None:
@@ -29,7 +31,7 @@ class pylint(Check):
             "--next",
             default=False,
             help="Next version of pylint is being tested.",
-            required=False,      
+            required=False,
         )
 
     def run(self, args: argparse.Namespace) -> int:
@@ -48,21 +50,34 @@ class pylint(Check):
             logger.info(f"Processing {package_name} for pylint check")
 
             # install dependencies
+            self.install_dev_reqs(executable, args, package_dir)
             try:
-                install_into_venv(executable, ["azure-pylint-guidelines-checker==0.5.6", "--index-url=https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/"])
+                install_into_venv(executable, ["azure-pylint-guidelines-checker==0.5.6", "--index-url=https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/"], package_dir)
             except CalledProcessError as e:
-                logger.error("Failed to install dependencies:", e)
+                logger.error(f"Failed to install dependencies: {e}")
                 return e.returncode
+
+            create_package_and_install(
+                distribution_directory=staging_directory,
+                target_setup=package_dir,
+                skip_install=False,
+                cache_dir=None,
+                work_dir=staging_directory,
+                force_create=False,
+                package_type="wheel",
+                pre_download_disabled=False,
+                python_executable=executable,
+            )
 
             # install pylint
             try:
                 if args.next:
                     # use latest version of pylint
-                    install_into_venv(executable, ["pylint"])
+                    install_into_venv(executable, ["pylint", f"PyGithub=={PYGITHUB_VERSION}"], package_dir)
                 else:
-                    install_into_venv(executable, [f"pylint=={PYLINT_VERSION}"])
+                    install_into_venv(executable, [f"pylint=={PYLINT_VERSION}"], package_dir)
             except CalledProcessError as e:
-                logger.error("Failed to install pylint:", e)
+                logger.error(f"Failed to install pylint: {e}")
                 return e.returncode
 
             top_level_module = parsed.namespace.split(".")[0]