+++ sdk/sql/azure-mgmt-sql/generated_samples/managed_ledger_digest_uploads_list.py
@@ -40,6 +40,6 @@ def main():
         print(item)
 
 
-# x-ms-original-file: specification/sql/resource-manager/Microsoft.Sql/preview/2024-11-01-preview/examples/ManagedLedgerDigestUploadsList.json
+# x-ms-original-file: specification/sql/resource-manager/Microsoft.Sql/preview/2022-08-01-preview/examples/ManagedLedgerDigestUploadsList.json
 if __name__ == "__main__":
     main()