+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Release History
 
-## 1.0.0b42 (Unreleased)
+## 1.0.0b43 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+- Fixes issues #42897 - TypeError in _transmit_from_storage: LocalFileBlob.get() returns None
+  ([#42897](https://github.com/Azure/azure-sdk-for-python/pull/42951))
+- Fix logic for message field on Microsoft.ApplicationInsights.Message to sync with Breeze
+  ([#43060](https://github.com/Azure/azure-sdk-for-python/pull/43060))
+
+### Other Changes
+
+## 1.0.0b42 (2025-09-17)
 
 ### Features Added
 - Customer Facing Statsbeat: Added remaining drop codes to base
@@ -19,10 +33,11 @@
   ([#42655](https://github.com/Azure/azure-sdk-for-python/pull/42655))
 - Customer Facing SDKStats: Added telemetry_success field to dropped items as per [Spec] - https://github.com/aep-health-and-standards/Telemetry-Collection-Spec/pull/606
   ([#42846](https://github.com/Azure/azure-sdk-for-python/pull/42846))
-
 ### Breaking Changes
 
 ### Bugs Fixed
+- Customer Facing SDKStats: Refactor to use `Manager` and `Singleton` pattern
+  ([#42969](https://github.com/Azure/azure-sdk-for-python/pull/42969))
 
 ### Other Changes